refactor lib
