+++ app/api/auth/password/route.ts
@@ -1,7 +1,7 @@
 import { NextResponse } from 'next/server';
 import { getServerSession } from 'next-auth/next';
 import bcrypt from 'bcryptjs';
-import dbConnect from '@/lib/mongoose';
+import dbConnect from '@/app/lib/mongoose';
 import User from '@/models/User';
 import { authOptions } from '../[...nextauth]/auth-options';
