+++ app/api/conversations/[id]/route.ts
@@ -1,6 +1,6 @@
 import { NextRequest, NextResponse } from 'next/server';
 import { getServerSession } from 'next-auth/next';
-import dbConnect from '@/lib/mongoose';
+import dbConnect from '@/app/lib/mongoose';
 import Conversation from '@/models/Conversation';
 import User from '@/models/User';
 import { authOptions } from '../../auth/[...nextauth]/auth-options';
