+++ app/api/create-checkout-session/route.ts
@@ -3,7 +3,7 @@ import Stripe from 'stripe';
 import { getServerSession } from 'next-auth';
 import { authOptions } from '../auth/[...nextauth]/auth-options';
 import User from '@/models/User';
-import dbConnect from '@/lib/mongoose';
+import dbConnect from '@/app/lib/mongoose';
 
 /**
  * Initialize Stripe instance with secret key
